refactor(uc-daemon): add uniclipd binary + decouple CLI from uc-desktop (ADR-008 P2)
Migrate all daemon host entry points, process bootstrap, and runtime
handles from uc-desktop into uc-daemon, then add a standalone uniclipd
binary target. The CLI now spawns uniclipd instead of re-invoking itself
with a hidden `daemon` subcommand, and no longer depends on uc-desktop.

Key changes:
- Move build_process_runtime / ProcessRuntimeContext into uc-daemon
- Move ProcessRuntimeHandles / start_in_process / run / run_standalone_from_env into uc-daemon
- Create DaemonProcessRuntime (standalone daemon counterpart to DesktopRuntime)
- Add [[bin]] name = "uniclipd" target to uc-daemon
- Move RUN_MODE_ENV / RUN_MODE_SERVER constants to uc-daemon-local/spawn_contract
- CLI spawns resolved uniclipd binary (sibling lookup + PATH fallback)
- Delete Commands::Daemon hidden subcommand
- Remove uc-desktop from uc-cli dependencies
- uc-desktop daemon/ and bootstrap.rs become pure re-export shims
