+++ src-tauri/crates/uc-daemon/Cargo.toml
@@ -3,11 +3,15 @@ name = "uc-daemon"
 version.workspace = true
 license.workspace = true
 edition = "2021"
-description = "GUI-agnostic daemon runtime extracted from uc-desktop (ADR-008 P1); hosts uniclipd later"
+description = "GUI-agnostic daemon runtime (ADR-008); hosts the uniclipd binary"
 
 [lib]
 name = "uc_daemon"
 path = "src/lib.rs"
+
+[[bin]]
+name = "uniclipd"
+path = "src/main.rs"
 
 [dependencies]
 uc-application = { path = "../uc-application" }
@@ -27,6 +31,7 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 chrono = { version = "0.4", features = ["serde"] }
 tracing = "0.1"
+rustls = { version = "0.23", default-features = false, features = ["ring", "std"] }
 
 [dev-dependencies]
 uc-daemon-local = { path = "../uc-daemon-local", features = ["test-helpers"] }
