+++ src-tauri/crates/uc-daemon/src/daemon/mod.rs
@@ -1,17 +1,19 @@
-//! daemon runtime 模块树（从 uc-desktop 迁出，ADR-008 P1）。
+//! daemon runtime 模块树（从 uc-desktop 迁出，ADR-008 P1/P2）。
 //!
 //! 保留 `daemon/` 路径使迁入文件内的 `crate::daemon::X` / `super::X` 引用
-//! 原样可解析。GuiInProcess 专属装配（`start_in_process` / `ProcessRuntimeHandles`）
-//! 仍在 uc-desktop 的 `host`，由其前向调用本模块的装配函数。
+//! 原样可解析。
 
 pub mod app;
 pub mod app_assembly;
 pub mod app_facade_assembly;
 pub mod bootstrap;
 pub mod handle;
+pub mod host;
 pub mod mobile_lan_lifecycle;
 pub mod ownership;
 pub mod peers;
+pub mod process_bootstrap;
+pub mod process_runtime;
 pub mod run_loop;
 pub mod run_mode;
 pub mod runtime_assembly;
